% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_cohorts_utils.R
\name{chrt_filter_select}
\alias{chrt_filter_select}
\title{Quickly perform filter & select routine used in \code{cohorts}}
\usage{
chrt_filter_select(
  x,
  ...,
  stayed = FALSE,
  served = FALSE,
  exited = FALSE,
  entered = FALSE,
  rm_dates,
  vars,
  start = rm_dates$calc$data_goes_back_to,
  end = rm_dates$meta_HUDCSV$Export_End,
  app_env = get_app_env(e = rlang::caller_env())
)
}
\arguments{
\item{x}{\code{(data.frame)} Must include columns necessary for the _*\emph{between} function used or any additional filter expressions supplied.}

\item{...}{\code{(expression)}s used to filter data}

\item{stayed}{\code{(logical)} Uses \link[HMIS]{stayed_between}}

\item{served}{\code{(logical)} Uses \link[HMIS]{served_between}}

\item{exited}{\code{(logical)} Uses \link[HMIS]{exited_between}}

\item{entered}{\code{(logical)} Uses \link[HMIS]{entered_between}}

\item{vars}{\code{(character)} vector of column names to retrain}

\item{start}{Default = ReportStart, character/date of the start of the date
range. Characters in format mdY, Ymd, dmY acceptable. Will be automatically
retrieved from parent environments if not specified. If start is named other
than ReportStart, it must be specified.}

\item{end}{Default = ReportEnd, character/date of the end of the date range.
Characters in format mdY, Ymd, dmY acceptable. Will be automatically
retrieved from parent environments if not specified. If end is named other
than ReportEnd, it must be specified.}

\item{app_env}{\code{(environment)} The application environment, defaults to the result of \code{get_app_env(e = rlang::caller_env())}.}
}
\value{
\code{(data.frame)} Output with filter and select applied.
}
\description{
All _*\emph{between} filters use \code{rm_dates$calc$data_goes_back_to} as the beginning of the period and \code{rm_dates$meta_HUDCSV$Export_End} as the end. See \code{dates} for details on these values.
}
