% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_cohorts_utils.R
\name{chrt_hoh_count}
\alias{chrt_hoh_count}
\title{Count number of households in the dataset. Used in \code{prep_cohorts}}
\usage{
chrt_hoh_count(served)
}
\arguments{
\item{served}{\code{(data.frame)}}
}
\value{
\code{(data.frame)} With a summary column using \link[dplyr]{n} named similarly to the input object name
}
\description{
Count number of households in the dataset. Used in \code{prep_cohorts}
}
