% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_client.R
\name{Client_add_UniqueID}
\alias{Client_add_UniqueID}
\title{Add the UniqueID to the Client export}
\usage{
Client_add_UniqueID(Client, Client_UniqueIDs)
}
\arguments{
\item{Client}{\code{(data.frame)} A data frame containing client information}

\item{Client_UniqueIDs}{\code{(data.frame)} A custom look linking PersonalID & UniqueID}

\item{app_env}{\code{(environment)} The application environment, defaults to the result of \code{get_app_env(e = rlang::caller_env())}.}
}
\value{
\code{(data.frame)} with UniqueID column
}
\description{
Add the UniqueID to the Client export
}
