% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{col_cats}
\alias{col_cats}
\title{Categories of HUD CSV Columns}
\format{
An object of class \code{list} of length 1.
}
\usage{
col_cats
}
\description{
Currently only groupings of Client columns
}
\keyword{datasets}
