% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_types}
\alias{data_types}
\title{FY 2022 HUD CSV Data groupings}
\format{
An object of class \code{list} of length 3.
}
\usage{
data_types
}
\description{
This list provides common groupings of specific data elements in the HUD CSV for ease of reference
}
\seealso{
destinations
}
\keyword{datasets}
