% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{destinations}
\alias{destinations}
\title{FY 2022 HUD CSV Data groupings}
\format{
An object of class \code{list} of length 4.
}
\usage{
destinations
}
\description{
This list provides common groupings of specific data elements in the HUD CSV for ease of reference
}
\keyword{datasets}
