% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_enrollment_utils.R
\name{Enrollment_add_AgeAtEntry_UniqueID}
\alias{Enrollment_add_AgeAtEntry_UniqueID}
\title{Add AgeAtEntry to Enrollment}
\usage{
Enrollment_add_AgeAtEntry_UniqueID(Enrollment, Client)
}
\value{
\code{(data.frame)} Enrollment with \code{AgeAtEntry} column
}
\description{
AgeAtEntry is the time elapsed from the Date of Birth \code{DOB} to the \code{EntryDate}
}
