% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_enrollment_utils.R
\name{Enrollment_add_HousingStatus}
\alias{Enrollment_add_HousingStatus}
\title{Add HousingStatus indicating the client's current housing status and Situation with details on that status.}
\usage{
Enrollment_add_HousingStatus(
  Enrollment_extra_Client_Exit_HH_CL_AaE,
  PH = data_types$Project$ProjectType$ph,
  Referrals = HMISdata::load_hmis_parquet("Referrals.parquet")
)
}
\arguments{
\item{Enrollment_extra_Client_Exit_HH_CL_AaE}{Custom Enrollment data frame}

\item{PH}{ProjectType codes considered Permanently Housed.  See \code{HMIS::hud_translations$ProjectType(table = TRUE)} & \code{?data_types}}

\item{Referrals}{Referrals data frame}

\item{prioritization_colors}{Custom data frame of prioritization colors}
}
\value{
Input data with HousingStatus & Situation column
}
\description{
Add HousingStatus indicating the client's current housing status and Situation with details on that status.
}
