% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_referrals.R
\name{filter_dupe_soft}
\alias{filter_dupe_soft}
\title{Filter duplicates without losing any values from \code{key}}
\usage{
filter_dupe_soft(.data, ..., key)
}
\arguments{
\item{.data}{\code{(data.frame)} Data with duplicates}

\item{...}{\code{(expressions)} filter expressions with which to filter}

\item{key}{\code{(name)} of the column key that will be grouped by and for which at least one observation will be preserved.}
}
\value{
\code{(data.frame)} without duplicates
}
\description{
Filter duplicates without losing any values from \code{key}
}
