% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dates.R
\name{load_dates}
\alias{load_dates}
\title{Load hard coded dates}
\usage{
load_dates(error = FALSE)
}
\arguments{
\item{error}{}
}
\value{
List of dates used in Rm apps
}
\description{
Load hard coded dates
}
