% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_program_lookup.R
\name{make_linked_df}
\alias{make_linked_df}
\title{Make UniqueID or EnrollmentID into a Clarity hyperlink}
\usage{
make_linked_df(
  .data,
  link_text,
  unlink = FALSE,
  new_ID,
  type = NULL,
  chr = TRUE
)
}
\arguments{
\item{.data}{\code{(data.frame)} The following columns are required for the specified link type:
\itemize{
\item{\code{PersonalID & UniqueID}}{ for Profile link}
\item{\code{PersonalID & EnrollmentID}}{ for Enrollment link}
\item{\code{PersonalID & EnrollmentID}}{ for Enrollment link}
}}

\item{link_text}{\code{(name)} unquoted of the column to unlink.}

\item{unlink}{\code{(logical)} Whether to turn the link back into the respective columns from which it was made.}

\item{new_ID}{\code{(name)} unquoted of the column to be created with the data from the linked column. (\code{PersonalID} will be recreated automatically if it doesn't exist).}
}
\value{
\code{(data.frame)} With \code{UniqueID} or \code{EnrollmentID} as a link
data.frame(a = letters, b = seq_along(letters)) |>  dplyr::mutate(a = make_link(a, b)) |> make_linked_df(a, unlink = TRUE)
}
\description{
Make UniqueID or EnrollmentID into a Clarity hyperlink
}
