% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_project.R
\name{pe_add_GrantType}
\alias{pe_add_GrantType}
\title{Add GrantType column to provider_extras}
\usage{
pe_add_GrantType(provider_extras)
}
\arguments{
\item{provider_extras}{\code{(data.frame)} provider_extras with Regions, see \code{pe_add_regions}.}
}
\value{
A data frame containing providers with numeric codes replaced.
}
\description{
GrantType indicates if the program is funded by one of HOPWA, PATH, SSVF, or RHY
}
