% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_project.R
\name{pe_add_ProjectType}
\alias{pe_add_ProjectType}
\title{Add ProjectType (dbl) to provider_extras}
\usage{
pe_add_ProjectType(provider_extras)
}
\arguments{
\item{provider_extras}{\code{(data.frame)} provider_extras with Regions, see \code{pe_add_regions}.}
}
\value{
\code{(data.frame)}
}
\description{
Add ProjectType (dbl) to provider_extras
}
