% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_project.R
\name{pe_create_APs}
\alias{pe_create_APs}
\title{Add Access Points to Provider_extras}
\usage{
pe_create_APs(provider_extras, ProjectCoC, dirs)
}
\arguments{
\item{provider_extras}{\code{(data.frame)} provider_extras with Regions, see \code{pe_add_regions}.}

\item{ProjectCoC}{\code{(data.frame)} Data frame containing information about the Project Continuum of Care (CoC).}

\item{dirs}{\code{(list)} List of directories used in the function.}
}
\value{
\code{(data.frame)} A data frame of Coordinated Entry Access Points with Counties & Populations Served.
}
\description{
Create a data frame of Coordinated Entry Access Points with information about the Counties & Populations Served.
}
