% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_enrollment.R
\name{prep_enrollment}
\alias{prep_enrollment}
\title{Load Enrollment as Enrollment_extra_Client_Exit_HH_CL_AaE}
\usage{
prep_enrollment(
  Enrollment,
  EnrollmentCoC,
  Enrollment_extras,
  Exit,
  Client = HMISdata::load_hmis_parquet("Client.parquet"),
  Project,
  rm_dates = load_dates()
)
}
\arguments{
\item{Enrollment_extras}{From Clarity Looker API Extras}
}
\description{
Load Enrollment as Enrollment_extra_Client_Exit_HH_CL_AaE
}
