% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_program_lookup.R
\name{prep_program_lookup}
\alias{prep_program_lookup}
\title{Fetch Program Lookup table}
\usage{
prep_program_lookup()
}
\arguments{
\item{program_lookup}{Custom program data frame}
}
\value{
\code{(tbl)}
}
\description{
Fetch Program Lookup table
}
