% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_project.R
\name{prep_project}
\alias{prep_project}
\title{Load Project & extras pre-requisites for \code{load_export}}
\usage{
prep_project()
}
\arguments{
\item{Regions}{From public data. See \code{load_public}}

\item{ProjectCoC}{From HUD CSV export}
}
\description{
Load Project & extras pre-requisites for \code{load_export}
}
