% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_referrals.R
\name{prep_referrals}
\alias{prep_referrals}
\title{Load Referrals}
\usage{
prep_referrals()
}
\description{
Loads Referrals, Referrals_full (unfiltered), and referral_result_summarize with filtering expressions used later
}
