% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_services.R
\name{prep_services}
\alias{prep_services}
\title{Load Services as Services_enroll_extras}
\usage{
prep_services()
}
\arguments{
\item{Enrollment_extra_Client_Exit_HH_CL_AaE}{Custom Enrollment data frame}
}
\description{
Load Services as Services_enroll_extras
}
