% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prioritization_colors}
\alias{prioritization_colors}
\title{Colors for the various level of housing status used in the prioritization report}
\format{
An object of class \code{character} of length 7.
}
\usage{
prioritization_colors
}
\description{
Colors for the various level of housing status used in the prioritization report
}
\keyword{datasets}
