% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_project.R
\name{Project_rm_zz}
\alias{Project_rm_zz}
\title{Filter Projects that have been retired (\code{zz}'d)}
\usage{
Project_rm_zz(Project)
}
\value{
A data frame containing projects that don't start with "zz".
}
\description{
Filter Projects that have been retired (\code{zz}'d)
}
