% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_client.R
\name{Client_redact}
\alias{Client_redact}
\title{Redact PII from Client HUD Export}
\usage{
Client_redact(Client)
}
\value{
\code{(tibble)} Redacted Client HUD Export
}
\description{
This redacts all PII (except DOB) from Client HUD Export
}
\concept{Client functions}
